+++ lib/arel/extensions.rb
@@ -27,6 +27,7 @@ require 'arel/extensions/any'
 require 'arel/extensions/array_subselect'
 require 'arel/extensions/type_cast'
 require 'arel/extensions/distinct_from'
+require 'arel/extensions/not_distinct_from'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
